Feat, Refactor: feat google auth with refactor etc codes

+++ domain/auth/auth.social-account.entity.go
@@ -3,9 +3,9 @@ package auth
 import "time"
 
 type SocialAccount struct {
-	ID        uint      `json:"id" db:"id"`
+	ID        int64     `json:"id" db:"id"`
 	UUID      string    `json:"uuid" db:"uuid"`
-	UserId    uint      `json:"user_id" db:"user_id"`
+	UserId    int64     `json:"user_id" db:"user_id"`
 	Provider  string    `json:"provider" db:"provider"`
 	SocialId  string    `json:"social_id" db:"social_id"`
 	CreatedAt time.Time `json:"created_at" db:"created_at"`
